+++ api/src/main/java/org/allaymc/api/eventbus/event/network/PacketEvent.java
@@ -10,6 +10,7 @@ import org.cloudburstmc.protocol.bedrock.packet.BedrockPacket;
 /**
  * @author daoge_cmd
  */
+// TODO: move to server
 @Getter
 @AllArgsConstructor
 public class PacketEvent extends NetworkEvent implements CancellableEvent {